Handle double-downloads in bisect-builds.py
When running this bisect command, to verify a bisect:

  tools\bisect-builds.py -g 991339 -b 991352 -a win64 --verify-range

an error was hit during cleanup. This happened because build 991339 was
downloaded twice, which happens because a --verify-range build downloads
three builds initially, and this range only contains two builds. The
same download gets deleted twice and the second deletion attempt fails.
This failure then causes a file to be left behind because the cleanup
process is halted. The error is:

    File "tools\bisect-builds.py", line 776, in Stop
      os.unlink(self.zip_file)
  FileNotFoundError: [WinError 2] The system cannot find the file specified: '991339-chrome-win32.zip'

The simplest fix is to make duplicate-deletion not an error. This
ensures that cleanup is done properly.

This error was hit on Windows. It is unknown whether it would be an
error on other operating systems.

Bug: 1430979
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4471227
Commit-Queue: Josiah Krutz <josiahk@google.com>
Auto-Submit: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1139279}

+++ tools/bisect-builds.py
@@ -791,7 +791,11 @@ class DownloadJob(object):
     assert self.thread, 'DownloadJob must be started before Stop is called.'
     self.quit_event.set()
     self.thread.join()
-    os.unlink(self.zip_file)
+    # If a very small range is requested, with verification, then the same
+    # download may be requested twice, so the second delete will fail. Check for
+    # that.
+    if os.path.exists(self.zip_file):
+      os.unlink(self.zip_file)
 
   def WaitFor(self):
     """Prints a message and waits for the download to complete. The download
